backend/requirements.txt
flask
pymongo
python-dotenv
flask-cors
